+++ lib/rdmm/queries/list_floors_query.rb
@@ -1,0 +1,6 @@
+module Rdmm
+  module Queries
+    class ListFloorsQuery < BaseQuery
+    end
+  end
+end