+++ lib/rdmm/resources/genre_resource.rb
@@ -1,0 +1,26 @@
+module Rdmm
+  module Resources
+    class GenreResource < BaseResource
+      # @return [Integer, nil]
+      def genre_id
+        source["genre_id"]
+      end
+
+      # @return [String, nil]
+      def list_url
+        source["list_url"]
+      end
+
+      # @return [String, nil]
+      def name
+        source["name"]
+      end
+
+      # @return [String, nil]
+      def ruby
+        source["ruby"]
+      end
+
+    end
+  end
+end