+++ lib/rdmm/responses/list_actresses_response.rb
@@ -1,0 +1,12 @@
+module Rdmm
+  module Responses
+    class ListActressesResponse < BaseResponse
+      # @note Override
+      def resources
+        body["result"]["actress"].map do |source|
+          ::Rdmm::Resources::ActressResource.new(source)
+        end
+      end
+    end
+  end
+end