+++ lib/rdmm/responses/list_authors_response.rb
@@ -1,0 +1,12 @@
+module Rdmm
+  module Responses
+    class ListAuthorsResponse < BaseResponse
+      # @note Override
+      def resources
+        (body["result"]["author"] || []).map do |source|
+          ::Rdmm::Resources::AuthorResource.new(source)
+        end
+      end
+    end
+  end
+end