+++ lib/rdmm/responses/list_makers_response.rb
@@ -1,0 +1,12 @@
+module Rdmm
+  module Responses
+    class ListMakersResponse < BaseResponse
+      # @note Override
+      def resources
+        body["result"]["maker"].map do |source|
+          ::Rdmm::Resources::MakerResource.new(source)
+        end
+      end
+    end
+  end
+end